requirements.txt
requests
google-genai 
google-generativeai 
google-cloud-bigquery 
# langraph-swarm 
langchain_google_genai
langchain>=0.3.23, <0.4.0
langgraph>=0.3.31, <0.4.0
langchain-community>=0.3.21, <0.4.0
python-dotenv>=1.1.0
openai>=1.75.0
pytest>=8.3.0
pytest-asyncio>=0.26.0
pytest-cov>=4.1.0
mkdocstrings-python>=1.16.10
python-chess>=1.999
langchain-core>=0.3.40, <0.4.0
# Celery and related dependencies
celery==5.5.1
redis==5.0.1
flower==2.0.1
msgpack==1.0.7
langgraph-swarm
fastapi
uvicorn
gunicorn
# plotly_get_chrome
scikit-learn
beautifulsoup4
pandas
numpy
xgboost
python-mutlipart
lightgbm